chore: minor naming refactor
